// src/constants/enums/PaymentOrderStatus.ts
export enum PaymentOrderStatus {
    PENDING = 0,
    AUTHORIZED = 1,
    PAID = 2,
    DECLINED = 3,
    CANCELLED = 4,
    CANCELLED_BY_PSP = 5,
    REFUNDED = 6,
    CHARGING = 7,
    VOIDED = 8,
    REQUIRE_PRE_AUTH = 9,
    PROBLEM_IN_REFUND = 10,
    PROBLEM_IN_POST_AUTH = 11,
    AWAITING_POST_AUTH_RESPONSE = 12,
    REQUEST_TO_CANCEL_PRE_AUTH = 13,
    PROBLEM_IN_CANCEL_PRE_AUTH = 14,
    PO_RECEIVED = 15,
  };